+++ test/support/factory.ex
@@ -547,9 +547,11 @@ defmodule TdDd.Factory do
   end
 
   def grant_request_group_factory(attrs) do
+    user_id = sequence(:user_id, &"#{&1}")
     %TdDd.Grants.GrantRequestGroup{
       id: sequence(:grant_request_group, &(&1 + 1_080)),
-      user_id: sequence(:user_id, &"#{&1}"),
+      user_id: user_id,
+      created_by_id: user_id,
       type: nil
     }
     |> merge_attributes(attrs)